requirements.txt
PyQt6>=6.6.0
anthropic>=0.25.0
Pillow>=10.0.0
keyboard>=0.13.5
winrt-Windows.Media.Ocr>=2.0.0
winrt-Windows.Graphics.Imaging>=2.0.0
winrt-Windows.Storage.Streams>=2.0.0
